Finished first pass of ImageCacheDisk, including freeing up space when necessary.

+++ app/src/main/java/com/barryholroyd/walmartproducts/ImageCacheMemory.java
@@ -66,7 +66,7 @@ class ImageCacheMemory extends BlobCacheMemory<String, Bitmap> {
      * @return    the size of the bitmap being stored.
      */
     @Override
-    protected int sizeOf(String key, Bitmap bm) {
+    protected long sizeOf(String key, Bitmap bm) {
         return bm.getByteCount();
     }
 }
